% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cond_qpc.R
\name{calc_cond_qpc}
\alias{calc_cond_qpc}
\title{Calculate Conditional Qpc}
\usage{
calc_cond_qpc(focal_trait, cor_trait, eigen_k, test_pcs, var_pcs)
}
\arguments{
\item{focal_trait}{vector of trait values for trait of interest. Not normalized yet.}

\item{cor_trait}{vector of trait values for the correlated trait. Not normalized yet}

\item{eigen_k}{named list from eigen decomposition of kinship matrix with 'vectors' and 'values'}

\item{test_pcs}{the vector of PCs you want to test for selection}

\item{var_pcs}{the vector of the PCs used to estimate Va}
}
\description{
This function calculates conditional Qpc given given using the eigen decomposition of the kinship matrix and trait values.
}
\examples{
calc_cond_qpc()
}
