% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_qpc.R
\name{calc_qpc}
\alias{calc_qpc}
\title{Calculate Qpc}
\usage{
calc_qpc(trait_values, eigen_k, test_pcs, var_pcs)
}
\arguments{
\item{trait_values}{vector of traits. Not normalized yet.}

\item{eigen_k}{Named list from eigen decomposition of kinship matrix with 'vectors' and 'values'}

\item{test_pcs}{the range of PCs you want to test for selection}

\item{var_pcs}{the range of PCs used to estimate Va}
}
\description{
This function calculates Qpc given using the eigen decomposition of the kinship matrix and trait values.
}
\examples{
calc_qpc()
}
