% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQpc.R
\name{calcQpc}
\alias{calcQpc}
\title{Calculate Qpc}
\usage{
calcQpc(myZ, myU, myLambdas, myL, myM)
}
\arguments{
\item{myZ}{vector of traits. Not normalized yet.}

\item{myU}{matrix of eigenvectors of the kinship matrix (each column is an eigenvector)}

\item{myLambdas}{vector of eigenvalues of the kinship matrix}

\item{myL}{the range of PCs used to estimate Va}

\item{myM}{the range of PCs you want to test for selection}
}
\description{
This function calculates Qpc given data about the relatedness matrix, and a set of trait values
}
\examples{
calcQpc()
}
