% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condQpc.R
\name{condQpc}
\alias{condQpc}
\title{Calculate conditional Qpc}
\usage{
condQpc(myZ, myU, myLambdas, myM, myL)
}
\arguments{
\item{myZ}{two columned data frame containing vectors of two traits. Not normalized yet. The first trait is the one we're testing for selection on conditional on the second trait.}

\item{myU}{matrix of eigenvectors of the kinship matrix (each column is an eigenvector)}

\item{myLambdas}{vector of eigenvalues of the kinship matrix}

\item{myM}{the range of PCs you want to test for selection}

\item{myL}{the range of PCs used to estimate Va}
}
\description{
This function calculates Qpc given data about the relatedness matrix, and a set of trait values
}
\examples{
calcQpc()
}
