% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cond_qpc.R
\name{plot_cond_qpc}
\alias{plot_cond_qpc}
\title{Plot Conditional Qpc results}
\usage{
plot_cond_qpc(cond_qpc, eigen_k, trait_values, test_pcs, pop_labels = NULL)
}
\arguments{
\item{cond_qpc}{the output of the calc_cond_qpc function}

\item{eigen_k}{Named list from eigen decomposition of kinship matrix with list items 'vectors' and 'values'}

\item{trait_values}{vector of focal trait values for each individual in the same order as eigen_k}

\item{test_pcs}{vector of tested PCs you want to plot}

\item{pop_labels}{optional vector of population labels for each individual in the same order as input for calc_qpc}
}
\description{
This function calculates Qpc given data about the relatedness matrix, and a set of trait values
}
\examples{
plot_cond_qpc()
}
