% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qpc.R
\name{plot_qpc}
\alias{plot_qpc}
\title{Plot Qpc results}
\usage{
plot_qpc(qpc, trait_values, eigen_k, test_pcs, pop_labels = NULL)
}
\arguments{
\item{qpc}{the output of the calc_qpc function.}

\item{trait_values}{vector of traits for each individual in the same order as input for calc_qpc}

\item{eigen_k}{named list from eigen decomposition of kinship matrix with list items 'vectors' and 'values'}

\item{test_pcs}{vector of tested PCs you want to plot}

\item{pop_labels}{optional vector of population labels for each individual in the same order as input for calc_qpc}
}
\description{
This function calculates Qpc given data about the relatedness matrix, and a set of trait values
}
\examples{
plot_qpc()
}
