% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaint_functions.R
\name{quaint}
\alias{quaint}
\title{Quaint for single pair}
\usage{
quaint(test_taxa, species_tree, quartet_table, outgroup = NULL, qt_vector)
}
\arguments{
\item{test_taxa}{A vector of taxa that you want to test for introgression between. Must be two taxa.}

\item{species_tree}{A species tree. The tip names must match those in the gene trees used to make the quartet table. Must be rooted correctly.}

\item{quartet_table}{A dataframe of quartet frequencies generated with the "quartetTable" or "quartetTableParallel" function from MSCquartets.}

\item{outgroup}{Optional but reccomended. The outgroup for your species tree. Can be one or more tip labels. If outgroup is not provided, quaint will run ABBA-BABA tests using all possible outgroups.}

\item{qt_vector}{Optional parameter that is used internally by the "quaint_all" function to speed up many tests.}
}
\value{
A dataframe with the results of each quaint ABBA-BABA test.
}
\description{
Run the quaint test for introgression on a single pair of taxa
}
\examples{
tbd
}
