% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaint_functions.R
\name{quaint_all}
\alias{quaint_all}
\title{Run quaint on every pair of taxa}
\usage{
quaint_all(species_tree, quartet_table, outgroup)
}
\arguments{
\item{species_tree}{A species tree. The tip names must match those in the gene trees used to make the quartet table. Must be rooted correctly.}

\item{quartet_table}{A dataframe of quartet frequencies generated with the "quartetTable" or "quartetTableParallel" function from MSCquartets.}

\item{outgroup}{Optional but reccomended. The outgroup for your species tree. Can be one or more tip labels. If outgroup is not provided, quaint will run ABBA-BABA tests using all possible outgroups.}
}
\value{
A dataframe with the results of each quaint ABBA-BABA test.
}
\description{
Run the quaint test for introgression on every possible pair of taxa in the species tree
}
\examples{
tbd
}
