% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaint_functions.R
\name{summarize_quaint_table}
\alias{summarize_quaint_table}
\title{Summarize quaint output table}
\usage{
summarize_quaint_table(quaint_table, alpha = 0.05, use_adjusted_p = TRUE)
}
\arguments{
\item{species_tree}{A species tree. The tip names must match those in the gene trees used to make the quartet table. Must be rooted correctly.}

\item{quartet_table}{A dataframe of quartet frequencies generated with the "quartetTable" or "quartetTableParallel" function from MSCquartets.}

\item{outgroup}{Optional but reccomended. The outgroup for your species tree. Can be one or more tip labels. If outgroup is not provided, quaint will run ABBA-BABA tests using all possible outgroups.}
}
\value{
A dataframe with a row for each pair of taxa and summary statistics of all ABBA-BABA tests used to test for introgression between them.
}
\description{
Summarize the output of quaint. Generates a dataframe with a row for each pair of taxa and summary statistics of all ABBA-BABA tests used to test for introgression between them.
}
\examples{
tbd
}
